Syntax improvements (#1)
Moved from https://github.com/zed-industries/zed/pull/22293.

- Fixes various issues with current syntax highlighting: modules and
patterns are not highlighted, number literals are treated as constants
etc. New highlights are mostly copied from
[nvim-treesitter](https://github.com/nvim-treesitter/nvim-treesitter/blob/master/queries/elm/highlights.scm)
with some modifications and removals of things that are not supported by
Zed. In particular, `@modules` are highlighted as `@title`, similarly to
how it's done in Zed's Haskell extension.
- Added `brackets.scm`.
- Changed the default tab_size to 4 as that's how most of Elm projects
are formatted and what elm-format uses.